refactor: user journey list
